Refactor truck GUI into modules
