+++ survey_cad_truck_gui/src/main.rs
@@ -4,7 +4,6 @@ use i_slint_common::sharedfontdb;
 use slint::platform::PointerEventButton;
 use slint::{Image, Model, PhysicalSize, SharedString, VecModel};
 use std::cell::RefCell;
-use std::io::Write;
 use std::rc::Rc;
 use std::time::Instant;
 
@@ -27,9 +26,8 @@ use survey_cad::styles::{
 use survey_cad::subassembly;
 use survey_cad::superelevation::SuperelevationPoint;
 mod snap;
-use serde::{Deserialize, Serialize};
 use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use truck_modeling::base::InnerSpace;
 use truck_modeling::base::Point3;
 use truck_modeling::base::Vector3;
@@ -42,8 +40,22 @@ mod truck_backend;
 use truck_backend::{HitObject, TruckBackend};
 mod persistence;
 use persistence::{load_layers, load_styles, save_layers, save_styles, StyleSettings};
+mod ui_state;
+mod commands;
+mod python;
+mod render;
 
 use once_cell::sync::Lazy;
+use commands::{Command, CommandStack, Context, MacroPlaying, MacroRecorder, apply_command, record_macro, spawn_line, spawn_point};
+use python::{MacroContext, PythonContext, play_macro_file, run_python_file};
+use render::{
+    WorkspaceRenderData, RenderState, RenderStyles, draw_text, screen_to_workspace,
+    workspace_to_screen, arc_from_three_points, arc_from_start_end_radius, polyline_to_solid,
+};
+use ui_state::{
+    Config, CursorFeedback, DragSelect, SnapPrefs, Vec2, DrawingMode, Theme,
+    WorkspaceProfile, load_config, save_config,
+};
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
 use rusttype::{point, Font, Scale};
@@ -60,736 +72,6 @@ static FONT_DATA: &[u8] = include_bytes!(concat!(
 ));
 static FONT: Lazy<Font<'static>> = Lazy::new(|| Font::try_from_bytes(FONT_DATA).unwrap());
 
-struct WorkspaceRenderData<'a> {
-    points: &'a [Point],
-    lines: &'a [(Point, Point)],
-    polygons: &'a [Vec<Point>],
-    polylines: &'a [Polyline],
-    arcs: &'a [Arc],
-    dimensions: &'a [LinearDimension],
-    surfaces: &'a [Tin],
-    alignments: &'a [Alignment],
-}
-
-#[derive(Default, Clone)]
-struct Vec2 {
-    x: f32,
-    y: f32,
-}
-
-#[derive(Default, Clone)]
-struct DragSelect {
-    start: (f32, f32),
-    end: (f32, f32),
-    active: bool,
-}
-
-#[derive(Default, Clone)]
-struct CursorFeedback {
-    pos: (f32, f32),
-    frame: u32,
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-struct SnapPrefs {
-    snap_to_grid: bool,
-    snap_to_entities: bool,
-    snap_points: bool,
-    snap_endpoints: bool,
-    snap_midpoints: bool,
-    snap_intersections: bool,
-    snap_nearest: bool,
-    snap_surfaces: bool,
-    snap_solids: bool,
-    snap_tolerance: f32,
-}
-
-impl Default for SnapPrefs {
-    fn default() -> Self {
-        Self {
-            snap_to_grid: true,
-            snap_to_entities: true,
-            snap_points: true,
-            snap_endpoints: true,
-            snap_midpoints: true,
-            snap_intersections: true,
-            snap_nearest: true,
-            snap_surfaces: true,
-            snap_solids: true,
-            snap_tolerance: 5.0,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Copy, Default)]
-enum WorkspaceProfile {
-    #[default]
-    Surveyor,
-    Engineer,
-    Gis,
-}
-
-#[derive(Serialize, Deserialize, Clone, Copy, Default)]
-enum Theme {
-    #[default]
-    Dark,
-    Light,
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-struct Config {
-    window_width: u32,
-    window_height: u32,
-    last_open_dir: Option<String>,
-    snap: SnapPrefs,
-    auto_tin: bool,
-    #[serde(default)]
-    quick_scripts: [String; 3],
-    #[serde(default)]
-    profile: WorkspaceProfile,
-    #[serde(default)]
-    theme: Theme,
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            window_width: 800,
-            window_height: 600,
-            last_open_dir: None,
-            snap: SnapPrefs::default(),
-            auto_tin: false,
-            quick_scripts: Default::default(),
-            profile: WorkspaceProfile::default(),
-            theme: Theme::default(),
-        }
-    }
-}
-
-fn config_path() -> Option<PathBuf> {
-    dirs::config_dir().map(|p| p.join("survey_cad_truck_gui").join("config.json"))
-}
-
-fn load_config() -> Config {
-    if let Some(path) = config_path() {
-        if let Ok(data) = fs::read_to_string(path) {
-            serde_json::from_str(&data).unwrap_or_default()
-        } else {
-            Config::default()
-        }
-    } else {
-        Config::default()
-    }
-}
-
-fn save_config(cfg: &Config) {
-    if let Some(path) = config_path() {
-        if let Some(parent) = path.parent() {
-            let _ = fs::create_dir_all(parent);
-        }
-        if let Ok(json) = serde_json::to_string_pretty(cfg) {
-            let _ = fs::write(path, json);
-        }
-    }
-}
-
-#[derive(Default, Clone, PartialEq)]
-enum DrawingMode {
-    #[default]
-    None,
-    Line {
-        start: Option<Point>,
-    },
-    Polygon {
-        vertices: Vec<Point>,
-    },
-    /// Center, start and end order
-    ArcCenter {
-        center: Option<Point>,
-        radius: Option<f64>,
-        start_angle: Option<f64>,
-    },
-    /// Three point arc
-    ArcThreePoint {
-        p1: Option<Point>,
-        p2: Option<Point>,
-    },
-    /// Start, end, then radius via third click
-    ArcStartEndRadius {
-        start: Option<Point>,
-        end: Option<Point>,
-        radius: Option<f64>,
-    },
-    Dimension {
-        start: Option<Point>,
-    },
-}
-
-#[derive(Clone)]
-enum Command {
-    RemovePoint {
-        index: usize,
-        point: Point,
-    },
-    AddPoint {
-        index: usize,
-        point: Point,
-    },
-    RemoveLine {
-        index: usize,
-        line: (Point, Point),
-    },
-    AddLine {
-        index: usize,
-        line: (Point, Point),
-    },
-    RemoveDimension {
-        index: usize,
-        dim: LinearDimension,
-    },
-    AddDimension {
-        index: usize,
-        dim: LinearDimension,
-    },
-    TinDeleteVertex {
-        surface: usize,
-        index: usize,
-        point: Point3,
-    },
-    TinAddVertex {
-        surface: usize,
-        index: usize,
-        point: Point3,
-    },
-}
-
-struct CommandStack {
-    undo: Vec<Command>,
-    redo: Vec<Command>,
-}
-
-#[derive(Default)]
-struct MacroRecorder {
-    file: Option<std::fs::File>,
-}
-
-#[derive(Default)]
-struct MacroPlaying(bool);
-
-fn record_macro(rec: &mut MacroRecorder, line: &str) {
-    if let Some(file) = &mut rec.file {
-        let _ = writeln!(file, "{line}");
-    }
-}
-
-impl CommandStack {
-    fn new() -> Self {
-        Self {
-            undo: Vec::new(),
-            redo: Vec::new(),
-        }
-    }
-
-    fn push(&mut self, cmd: Command) {
-        self.undo.push(cmd);
-        self.redo.clear();
-    }
-
-    fn undo(&mut self, ctx: &Context) {
-        if let Some(cmd) = self.undo.pop() {
-            let inverse = apply_command(&cmd, ctx);
-            self.redo.push(inverse);
-        }
-    }
-
-    fn redo(&mut self, ctx: &Context) {
-        if let Some(cmd) = self.redo.pop() {
-            let inverse = apply_command(&cmd, ctx);
-            self.undo.push(inverse);
-        }
-    }
-}
-
-struct Context<'a> {
-    points: &'a Rc<RefCell<PointDatabase>>,
-    point_styles: &'a Rc<RefCell<Vec<usize>>>,
-    lines: &'a Rc<RefCell<Vec<(Point, Point)>>>,
-    line_styles: &'a Rc<RefCell<Vec<usize>>>,
-    dimensions: &'a Rc<RefCell<Vec<LinearDimension>>>,
-    backend: &'a Rc<RefCell<TruckBackend>>,
-}
-
-fn apply_command(cmd: &Command, ctx: &Context) -> Command {
-    match cmd {
-        Command::RemovePoint { index, point } => {
-            ctx.points.borrow_mut().remove(*index);
-            ctx.point_styles.borrow_mut().remove(*index);
-            ctx.backend.borrow_mut().remove_point(*index);
-            Command::AddPoint {
-                index: *index,
-                point: *point,
-            }
-        }
-        Command::AddPoint { index, point } => {
-            ctx.points.borrow_mut().insert(*index, *point);
-            ctx.point_styles.borrow_mut().insert(*index, 0);
-            ctx.backend.borrow_mut().add_point(point.x, point.y, 0.0);
-            Command::RemovePoint {
-                index: *index,
-                point: *point,
-            }
-        }
-        Command::RemoveLine { index, line } => {
-            ctx.lines.borrow_mut().remove(*index);
-            ctx.line_styles.borrow_mut().remove(*index);
-            ctx.backend.borrow_mut().remove_line(*index);
-            Command::AddLine {
-                index: *index,
-                line: *line,
-            }
-        }
-        Command::AddLine { index, line } => {
-            ctx.lines.borrow_mut().insert(*index, *line);
-            ctx.line_styles.borrow_mut().insert(*index, 0);
-            ctx.backend.borrow_mut().add_line(
-                [line.0.x, line.0.y, 0.0],
-                [line.1.x, line.1.y, 0.0],
-                [1.0, 1.0, 1.0, 1.0],
-                1.0,
-            );
-            Command::RemoveLine {
-                index: *index,
-                line: *line,
-            }
-        }
-        Command::RemoveDimension { index, dim } => {
-            ctx.backend.borrow_mut().remove_dimension(*index);
-            ctx.dimensions.borrow_mut().remove(*index);
-            Command::AddDimension {
-                index: *index,
-                dim: dim.clone(),
-            }
-        }
-        Command::AddDimension { index, dim } => {
-            ctx.dimensions.borrow_mut().insert(*index, dim.clone());
-            ctx.backend.borrow_mut().add_dimension(
-                [dim.start.x, dim.start.y, 0.0],
-                [dim.end.x, dim.end.y, 0.0],
-                [1.0, 1.0, 1.0, 1.0],
-                1.0,
-            );
-            Command::RemoveDimension {
-                index: *index,
-                dim: dim.clone(),
-            }
-        }
-        Command::TinDeleteVertex {
-            surface,
-            index,
-            point,
-        } => {
-            ctx.backend.borrow_mut().delete_vertex(*surface, *index);
-            Command::TinAddVertex {
-                surface: *surface,
-                index: *index,
-                point: *point,
-            }
-        }
-        Command::TinAddVertex {
-            surface,
-            index,
-            point,
-        } => {
-            ctx.backend.borrow_mut().add_vertex(*surface, *point);
-            Command::TinDeleteVertex {
-                surface: *surface,
-                index: *index,
-                point: *point,
-            }
-        }
-    }
-}
-
-struct RenderState<'a> {
-    offset: &'a Rc<RefCell<Vec2>>,
-    zoom: &'a Rc<RefCell<f32>>,
-    selected: &'a Rc<RefCell<Vec<usize>>>,
-    selected_lines: &'a Rc<RefCell<Vec<(Point, Point)>>>,
-    selected_polygons: &'a Rc<RefCell<Vec<usize>>>,
-    selected_polylines: &'a Rc<RefCell<Vec<usize>>>,
-    selected_arcs: &'a Rc<RefCell<Vec<usize>>>,
-    selected_dimensions: &'a Rc<RefCell<Vec<usize>>>,
-    drag: &'a Rc<RefCell<DragSelect>>,
-    cursor_feedback: &'a Rc<RefCell<Option<CursorFeedback>>>,
-    snap_target: &'a Rc<RefCell<Option<Point>>>,
-}
-
-struct RenderStyles<'a> {
-    point_styles: &'a [PointStyle],
-    style_indices: &'a Rc<RefCell<Vec<usize>>>,
-    line_styles: &'a [LineStyle],
-    line_style_indices: &'a Rc<RefCell<Vec<usize>>>,
-    polygon_styles: &'a [survey_cad::styles::PolygonStyle],
-    polygon_style_indices: &'a Rc<RefCell<Vec<usize>>>,
-    alignment_style: &'a LineStyle,
-    show_labels: bool,
-    label_style: &'a LineLabelStyle,
-    point_label_style: &'a PointLabelStyle,
-    show_point_numbers: bool,
-}
-
-fn draw_text(
-    pixmap: &mut Pixmap,
-    text: &str,
-    font: &Font,
-    x: f32,
-    y: f32,
-    color: Color,
-    size: f32,
-) {
-    let scale = Scale::uniform(size);
-    let v_metrics = font.v_metrics(scale);
-    let mut cursor = x;
-    for ch in text.chars() {
-        let glyph = font
-            .glyph(ch)
-            .scaled(scale)
-            .positioned(point(cursor, y + v_metrics.ascent));
-        if let Some(bb) = glyph.pixel_bounding_box() {
-            glyph.draw(|gx, gy, gv| {
-                let px = gx as i32 + bb.min.x;
-                let py = gy as i32 + bb.min.y;
-                if px >= 0
-                    && py >= 0
-                    && (px as u32) < pixmap.width()
-                    && (py as u32) < pixmap.height()
-                {
-                    let idx = (py as u32 * pixmap.width() + px as u32) as usize;
-                    pixmap.pixels_mut()[idx] = tiny_skia::ColorU8::from_rgba(
-                        (color.red() * 255.0) as u8,
-                        (color.green() * 255.0) as u8,
-                        (color.blue() * 255.0) as u8,
-                        (gv * 255.0) as u8,
-                    )
-                    .premultiply();
-                }
-            });
-        }
-        cursor += glyph.unpositioned().h_metrics().advance_width;
-    }
-}
-
-fn screen_to_workspace(
-    x: f32,
-    y: f32,
-    offset: &Rc<RefCell<Vec2>>,
-    zoom: &Rc<RefCell<f32>>,
-    width: f32,
-    height: f32,
-) -> Point {
-    let origin_x = width / 2.0;
-    let origin_y = height / 2.0;
-    let z = *zoom.borrow();
-    let off = offset.borrow();
-    let wx = (x - origin_x) / z - off.x;
-    let wy = -((y - origin_y) / z) - off.y;
-    Point::new(wx as f64, wy as f64)
-}
-
-fn workspace_to_screen(
-    p: Point,
-    offset: &Rc<RefCell<Vec2>>,
-    zoom: &Rc<RefCell<f32>>,
-    width: f32,
-    height: f32,
-) -> (f32, f32) {
-    let origin_x = width / 2.0;
-    let origin_y = height / 2.0;
-    let z = *zoom.borrow();
-    let off = offset.borrow();
-    let sx = (p.x as f32 + off.x) * z + origin_x;
-    let sy = origin_y - (p.y as f32 + off.y) * z;
-    (sx, sy)
-}
-
-fn arc_from_three_points(p1: Point, p2: Point, p3: Point) -> Option<Arc> {
-    let a = p2.x - p1.x;
-    let b = p2.y - p1.y;
-    let c = p3.x - p1.x;
-    let d = p3.y - p1.y;
-    let e = a * (p1.x + p2.x) + b * (p1.y + p2.y);
-    let f = c * (p1.x + p3.x) + d * (p1.y + p3.y);
-    let g = 2.0 * (a * (p3.y - p2.y) - b * (p3.x - p2.x));
-    if g.abs() < f64::EPSILON {
-        return None;
-    }
-    let cx = (d * e - b * f) / g;
-    let cy = (a * f - c * e) / g;
-    let center = Point::new(cx, cy);
-    let r = ((center.x - p1.x).powi(2) + (center.y - p1.y).powi(2)).sqrt();
-    let sa = (p1.y - cy).atan2(p1.x - cx);
-    let mut ma = (p2.y - cy).atan2(p2.x - cx);
-    let mut ea = (p3.y - cy).atan2(p3.x - cx);
-    let cross = (p2.x - p1.x) * (p3.y - p2.y) - (p2.y - p1.y) * (p3.x - p2.x);
-    if cross >= 0.0 {
-        while ma < sa {
-            ma += 2.0 * std::f64::consts::PI;
-        }
-        while ea < ma {
-            ea += 2.0 * std::f64::consts::PI;
-        }
-    } else {
-        while ma > sa {
-            ma -= 2.0 * std::f64::consts::PI;
-        }
-        while ea > ma {
-            ea -= 2.0 * std::f64::consts::PI;
-        }
-    }
-    Some(Arc::new(center, r, sa, ea))
-}
-
-fn arc_from_start_end_radius(start: Point, end: Point, r: f64, orient: Point) -> Option<Arc> {
-    let dx = end.x - start.x;
-    let dy = end.y - start.y;
-    let chord = (dx * dx + dy * dy).sqrt();
-    if r <= chord / 2.0 {
-        return None;
-    }
-    let mid = Point::new((start.x + end.x) / 2.0, (start.y + end.y) / 2.0);
-    let len = (dx * dx + dy * dy).sqrt();
-    if len.abs() < f64::EPSILON {
-        return None;
-    }
-    let perp = (-dy / len, dx / len);
-    let h = (r * r - (chord / 2.0).powi(2)).sqrt();
-    let sign = ((orient.x - start.x) * dy - (orient.y - start.y) * dx).signum();
-    let cx = mid.x + perp.0 * h * sign;
-    let cy = mid.y + perp.1 * h * sign;
-    let center = Point::new(cx, cy);
-    let sa = (start.y - cy).atan2(start.x - cx);
-    let ea = (end.y - cy).atan2(end.x - cx);
-    Some(Arc::new(center, r, sa, ea))
-}
-
-fn spawn_point(
-    points: &Rc<RefCell<PointDatabase>>,
-    styles: &Rc<RefCell<Vec<usize>>>,
-    backend: &Rc<RefCell<TruckBackend>>,
-    p: Point,
-) {
-    points.borrow_mut().push(p);
-    styles.borrow_mut().push(0);
-    backend.borrow_mut().add_point(p.x, p.y, 0.0);
-}
-
-fn spawn_line(
-    points: &Rc<RefCell<PointDatabase>>,
-    lines: &Rc<RefCell<Vec<(Point, Point)>>>,
-    point_styles: &Rc<RefCell<Vec<usize>>>,
-    line_styles: &Rc<RefCell<Vec<usize>>>,
-    backend: &Rc<RefCell<TruckBackend>>,
-    a: Point,
-    b: Point,
-) {
-    spawn_point(points, point_styles, backend, a);
-    spawn_point(points, point_styles, backend, b);
-    lines.borrow_mut().push((a, b));
-    line_styles.borrow_mut().push(0);
-    backend
-        .borrow_mut()
-        .add_line([a.x, a.y, 0.0], [b.x, b.y, 0.0], [1.0, 1.0, 1.0, 1.0], 1.0);
-}
-
-struct MacroContext {
-    playing: Rc<RefCell<MacroPlaying>>,
-    recorder: Rc<RefCell<MacroRecorder>>,
-    point_db: Rc<RefCell<PointDatabase>>,
-    point_styles: Rc<RefCell<Vec<usize>>>,
-    lines: Rc<RefCell<Vec<(Point, Point)>>>,
-    line_styles: Rc<RefCell<Vec<usize>>>,
-    backend: Rc<RefCell<TruckBackend>>,
-    render_image: Rc<dyn Fn() -> Image>,
-    weak: slint::Weak<MainWindow>,
-}
-
-struct PythonContext {
-    weak: slint::Weak<MainWindow>,
-    point_db: Rc<RefCell<PointDatabase>>,
-    lines: Rc<RefCell<Vec<(Point, Point)>>>,
-    surfaces: Rc<RefCell<Vec<Tin>>>,
-    selected_points: Rc<RefCell<Vec<usize>>>,
-    selected_lines: Rc<RefCell<Vec<(Point, Point)>>>,
-    offset: Rc<RefCell<Vec2>>,
-    zoom: Rc<RefCell<f32>>,
-}
-
-fn play_macro_file(path: &Path, ctx: &MacroContext) {
-    if let Ok(content) = std::fs::read_to_string(path) {
-        ctx.playing.borrow_mut().0 = true;
-        for line in content.lines() {
-            let parts = shell_words::split(line).unwrap_or_default();
-            if parts.is_empty() {
-                continue;
-            }
-            match parts[0].as_str() {
-                "point" if parts.len() >= 3 => {
-                    if let (Ok(x), Ok(y)) = (parts[1].parse::<f64>(), parts[2].parse::<f64>()) {
-                        spawn_point(
-                            &ctx.point_db,
-                            &ctx.point_styles,
-                            &ctx.backend,
-                            Point::new(x, y),
-                        );
-                    }
-                }
-                "line" if parts.len() >= 5 => {
-                    if let (Ok(x1), Ok(y1), Ok(x2), Ok(y2)) = (
-                        parts[1].parse::<f64>(),
-                        parts[2].parse::<f64>(),
-                        parts[3].parse::<f64>(),
-                        parts[4].parse::<f64>(),
-                    ) {
-                        spawn_line(
-                            &ctx.point_db,
-                            &ctx.lines,
-                            &ctx.point_styles,
-                            &ctx.line_styles,
-                            &ctx.backend,
-                            Point::new(x1, y1),
-                            Point::new(x2, y2),
-                        );
-                    }
-                }
-                _ => {}
-            }
-        }
-        ctx.playing.borrow_mut().0 = false;
-        ctx.recorder.borrow_mut().file = None;
-        if let Some(app) = ctx.weak.upgrade() {
-            if app.get_workspace_mode() == 0 {
-                app.set_workspace_image((*ctx.render_image)());
-                app.window().request_redraw();
-            }
-            refresh_workspace(&app, &*ctx.render_image, &ctx.backend);
-        }
-    }
-}
-
-fn run_python_file(path: &Path, ctx: &PythonContext) {
-    match std::fs::read_to_string(path) {
-        Ok(code) => {
-            let result = Python::with_gil(|py| {
-                let module = PyModule::new_bound(py, "survey_cad_python")?;
-                survey_cad_python::init(py, &module)?;
-
-                let pts: Vec<Py<survey_cad_python::Point>> = ctx
-                    .point_db
-                    .borrow()
-                    .iter()
-                    .map(|p| Py::new(py, survey_cad_python::Point::new(p.x, p.y)))
-                    .collect::<PyResult<_>>()?;
-
-                let lines_py: Vec<(Py<survey_cad_python::Point>, Py<survey_cad_python::Point>)> =
-                    ctx.lines
-                        .borrow()
-                        .iter()
-                        .map(|(a, b)| {
-                            Ok((
-                                Py::new(py, survey_cad_python::Point::new(a.x, a.y))?,
-                                Py::new(py, survey_cad_python::Point::new(b.x, b.y))?,
-                            ))
-                        })
-                        .collect::<PyResult<_>>()?;
-
-                let surfs: Vec<Py<PyAny>> = ctx
-                    .surfaces
-                    .borrow()
-                    .iter()
-                    .map(|s| {
-                        let dict = PyDict::new_bound(py);
-                        let verts: Vec<(f64, f64, f64)> =
-                            s.vertices.iter().map(|v| (v.x, v.y, v.z)).collect();
-                        let tris: Vec<(usize, usize, usize)> =
-                            s.triangles.iter().map(|t| (t[0], t[1], t[2])).collect();
-                        dict.set_item("vertices", verts)?;
-                        dict.set_item("triangles", tris)?;
-                        Ok(dict.into())
-                    })
-                    .collect::<PyResult<_>>()?;
-
-                let selected_pts: Vec<usize> = ctx.selected_points.borrow().clone();
-                let selected_lines_py: Vec<(
-                    Py<survey_cad_python::Point>,
-                    Py<survey_cad_python::Point>,
-                )> = ctx
-                    .selected_lines
-                    .borrow()
-                    .iter()
-                    .map(|(a, b)| {
-                        Ok((
-                            Py::new(py, survey_cad_python::Point::new(a.x, a.y))?,
-                            Py::new(py, survey_cad_python::Point::new(b.x, b.y))?,
-                        ))
-                    })
-                    .collect::<PyResult<_>>()?;
-
-                let view = PyDict::new_bound(py);
-                let off = ctx.offset.borrow();
-                view.set_item("offset", (off.x, off.y))?;
-                view.set_item("zoom", *ctx.zoom.borrow())?;
-
-                let globals = PyDict::new_bound(py);
-                globals.set_item("survey_cad_python", module)?;
-                globals.set_item("points", pts)?;
-                globals.set_item("lines", lines_py)?;
-                globals.set_item("surfaces", surfs)?;
-                globals.set_item("selected_points", selected_pts)?;
-                globals.set_item("selected_lines", selected_lines_py)?;
-                globals.set_item("view", view)?;
-
-                py.run_bound(&code, Some(&globals), None)
-            });
-
-            match result {
-                Ok(_) => {
-                    if let Some(app) = ctx.weak.upgrade() {
-                        app.set_status(SharedString::from("Python script finished"));
-                    }
-                }
-                Err(e) => {
-                    if let Some(app) = ctx.weak.upgrade() {
-                        app.set_status(SharedString::from(format!("Python error: {e}")));
-                    }
-                }
-            }
-        }
-        Err(e) => {
-            if let Some(app) = ctx.weak.upgrade() {
-                app.set_status(SharedString::from(format!("Failed to read: {e}")));
-            }
-        }
-    }
-}
-
-fn polyline_to_solid(pl: &Polyline, vector: Vector3) -> Option<Solid> {
-    if pl.vertices.len() < 3 {
-        return None;
-    }
-    let verts: Vec<_> = pl
-        .vertices
-        .iter()
-        .map(|p| builder::vertex(Point3::new(p.x, p.y, 0.0)))
-        .collect();
-    let mut edges = Vec::new();
-    for i in 0..verts.len() {
-        edges.push(builder::line(&verts[i], &verts[(i + 1) % verts.len()]));
-    }
-    let wire = Wire::from_iter(edges);
-    let face = builder::try_attach_plane(&[wire]).ok()?;
-    let solid: Solid = builder::tsweep(&face, vector);
-    Some(solid)
-}
 
 fn render_workspace(
     data: &WorkspaceRenderData,
